updates to embeds and purge command

+++ commands/mod/purgs.js
@@ -1,5 +1,11 @@
 const { Command } = require('discord.js-commando');
 const oneline = require('common-tags').oneLine;
+const { TagDB } = require('../../mysql.js');
+const config = require('../../config.json');
+const SqlTag = new TagDB(config.tagHost, config.tagUser, config.tagPassword, config.caseDatabase);
+const sql = SqlTag.getConnection();
+const Embed = require('../../embeds.js');
+const winston = require('winston');
 
 module.exports = class PurgeCommand extends Command {
 	constructor(client) {
@@ -10,7 +16,7 @@ module.exports = class PurgeCommand extends Command {
 			description: oneline`
 						Purgt eine bestimmte Anzahl an Nachrichten.
 						Es kann ein User oder String angegeben werden um nur Nachrichten von diesem User oder mit dem angegeben String zu löschen.`,
-			format: '<user oder String>',
+			format: '<Messages to purge> <reason> <user>',
 
 			args: [
 				{
@@ -18,6 +24,12 @@ module.exports = class PurgeCommand extends Command {
 					prompt: 'Bitte gebe eine Anzahl von Nachrichten an, die gelöscht werden sollen',
 					type: 'integer',
 					max: 100
+				},
+				{
+					key: 'reason',
+					prompt: 'Gebe einen Grund für den Purge an (Nicht mehr bearbeitbar!)',
+					type: 'string',
+					max: 200
 				},
 				{
 					key: 'user',
@@ -38,10 +50,19 @@ module.exports = class PurgeCommand extends Command {
 		let limit = args.limit + 1;
 		let user = args.user.id;
 		let query = args.content;
+		let reason = args.reason;
+		let caseNumber;
+		if (!msg.member.hasPermission('MANAGE_MESSAGES')) return msg.reply('du hast keine Rechte Nachrichten zu löschen');
+		if (!msg.guild.member(this.client.user.id).hasPermission('MANAGE_MESSAGES')) return msg.reply('Ich habe keine Rechte Nachrichten zu Löschen');
 		if (user.length > 0 && query.length > 0) {
 			msg.reply('Du kannst nur entweder ein User oder eine Nachricht eingeben');
 			return;
 		}
+		sql.query('INSERT INTO `cases`(`caseAction`, `caseModerator`,`caseReason`,`caseMessageID`, `caseChannel`, `caseLimit`) VALUES ("Purge", ?, ?, ?, ?, ?)', [msg.author.id, reason.length === 0 ? 'None' : reason, 'Placeholder', msg.channel.id, limit], (err, results) => {
+			if (err) return msg.reply(`Ups... Querry Error: ${err}`);
+			caseNumber = results.insertId;
+			this.message(caseNumber, msg);
+		});
 		if (user.length > 0) return this.purgeUser(msg, user, limit);
 		if (query.length > 0) return this.purgeQuery(msg, query, limit);
 		return this.purge(msg, limit);
@@ -52,6 +73,7 @@ module.exports = class PurgeCommand extends Command {
 			let filtered = messages.filter(m => m.author.id === user);
 			msg.channel.bulkDelete(filtered);
 		});
+		return;
 	}
 
 	async purgeQuery(msg, query, limit) {
@@ -60,10 +82,38 @@ module.exports = class PurgeCommand extends Command {
 			let filtered = messages.filter(m => m.content.toLowerCase() === query.toLowerCase());
 			msg.channel.bulkDelete(filtered);
 		});
+		return;
 	}
 
-	async purge(msg, args) {
+	async purge(msg, args, limit) {
+		msg.channel.fetchMessages({ limit: limit })
+		.then(messages => {
+			msg.channel.bulkDelete(messages);
+		});
+		return;
+	}
+	async message(caseNumber, msg) {
+		sql.query('SELECT * FROM `cases` WHERE `caseNumber` = ?', [caseNumber], (err, results) => {
+			if (err) msg.channel.sendMessage(`Ups... Error getting Query for Modlog ${err}`);
+			let caseNum = results[0].caseNumber;
+			let action = results[0].caseAction;
+			let ModID = results[0].caseModerator;
+			let reason = results[0].caseReason;
+			let mod = msg.guild.members.get(ModID);
+			let channel = msg.guild.channels.get(results[0].caseChannel);
+			let channelName = channel.name;
+			let limit = results[0].caseLimit;
+			const newEmbed = new Embed(this.client, msg, caseNum, action, null, mod, reason, null, limit, channelName);
+			let embed = newEmbed.purgeCase();
 
+			let modChannel = msg.guild.channels.find('name', 'logtest');
+			modChannel.sendMessage('', { embed })
+			.then(message => {
+				sql.query('UPDATE `cases` SET `caseMessageID` = ? WHERE `caseNumber` = ?', [message.id, caseNum], (er, res) => {
+					if (err) return msg.reply(`Ups es ist ein Fehler Aufgetreten. ERROR ADDING MESSAGE ID TO CASE : ${er} ${res}`);
+				});
+			}).catch(e => winston.error(e));
+		});
 	}
 };
 // fuck this fucking fuck I'll do that later
